change: create FolderService
* move Folder-related methods from CapacityService to FolderService
